Refactor BaselineDomainScanner to use DomainTableId enum
Refactored `BaselineDomainScanner` to use a strongly-typed `DomainTableId` enum instead of hardcoded integers for domain IDs. This improves code readability and maintainability.

- Created `DomainTableId` enum in `DHSIntegrationAgent.Domain.Claims`.
- Updated `BaselineDomainScanner.cs` to use the new enum.
- Updated `FetchStageService.cs` and `DispatchService.cs` to handle the enum, casting to int where necessary for persistence compatibility.
- Verified with `dotnet build` and unit tests.

+++ src/DHSIntegrationAgent.Workers/DispatchService.cs
@@ -407,7 +407,7 @@ public sealed class DispatchService : IDispatchService
                 {
                     if (_domainLookupByName.TryGetValue(f.DomainName, out var domain))
                     {
-                        if (mappingLookup.TryGetValue((domain.DomainTableId, valLower), out var mapping))
+                        if (mappingLookup.TryGetValue(((int)domain.DomainTableId, valLower), out var mapping))
                         {
                             enrichments ??= new();
                             enrichments.Add((f.TargetField, new JsonObject
